+++ wallet-chain-interact/src/tron/operations/stake/delegate.rs
@@ -84,8 +84,8 @@ impl TronTxOperation<DelegateResp> for DelegateArgs {
         self.receiver_address.clone()
     }
 
-    fn get_value(&self) -> i64 {
-        self.balance / consts::TRX_VALUE
+    fn get_value(&self) -> f64 {
+        (self.balance / consts::TRX_VALUE) as f64
     }
 }
